==> analysis/analyses/3-score_rules/1-generate-toys.Rmd <==
---
title: "1-generate-toys"
output: html_document
---

```{r setup}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
library(dynalysis)
library(tidyverse)
library(dyneval)
```

## Generate toy datasets for comparing scores
```{r}
source("analysis/analyses/3-score_rules/generation.R")
source("analysis/analyses/3-score_rules/perturbation.R")

experiment("score_rules")
```


```{r}
toys_blueprint <- tribble(
  ~trajectory_model, ~perturbator_id,
  "simple_linear", "gs",
  "simple_linear", "switch_two_cells",
  "simple_linear", "switch_all_cells",
  "simple_linear", "join_linear",
  "simple_linear", "split_linear",
  "simple_linear", "warp",
  "simple_linear", "hairy",
  "simple_linear", "hairy_small",
  "simple_linear", "hairy_large",
  "simple_linear", "remove_cells",
  "simple_linear", "add_distant_edge",
  "simple_linear", "structure_and_position",
  "bifurcating", "gs",
  "bifurcating", "change_lengths",
  "bifurcating", "change_terminal_lengths",
  "bifurcating", "switch_two_cells",
  "bifurcating", "switch_all_cells",
  "bifurcating", "warp",
  "bifurcating", "hairy",
  "bifurcating", "remove_cells",
  "bifurcating", "add_distant_edge",
  "bifurcating", "structure_and_position",
  "cycle", "gs",
  "cycle", "switch_all_cells",
  "cycle", "break_cycles",
  "cycle", "warp",
  "cycle", "hairy",
  "cycle", "remove_cells",
  "cycle", "add_distant_edge"
) %>% 
  bind_rows(tibble(trajectory_model="simple_linear", perturbator_id=paste0("switch_", percs * 100))) %>%
  bind_rows(crossing(trajectory_model=trajectory_models, perturbator_id=paste0("change_network_", trajectory_models)) %>% mutate(ncells = 5)) %>% 
  rowwise() %>% 
  mutate(
  perturbator=list(get(paste0("perturb_", perturbator_id)))
)

# replicate
nreplicates <- 1
toys <- toys_blueprint %>% slice(rep(1:n(), each=nreplicates)) %>% mutate(
  replicate=seq_len(nrow(.))%%nreplicates,
  toy_category=paste0(trajectory_model, "-", perturbator_id),
  toy_id=paste0(toy_category, "-", replicate)
) %>%
  mutate(ncells=sample(50:500, n(), replace=TRUE))

# generate gold standards and toys, can take some time (for computing the geodesic distances I presume)
# I choose to not do this using mutate because it is much easier to debug using loops
toys$gs <- toys %>% split(seq_len(nrow(toys))) %>% parallel::mclapply(function(row) {
  dyntoy:::generate_dataset("why", row$trajectory_model, row$ncells)
}, mc.cores=8)
#toys$gs <- map2(toys$generator, toys$ncells, ~.x(.y))

toys$toy <- toys %>% split(seq_len(nrow(toys))) %>% parallel::mclapply(function(row) {
  row$perturbator[[1]](row$gs[[1]])
}, mc.cores=8)
# toys$toy <- map2(toys$perturbator, toys$gs, ~.x(.y))
toys$toy <- map2(toys$toy, toys$toy_id, ~rename_toy(.x, .y))

toys$gs_grouped <- parallel::mclapply(toys$gs, group_task, mc.cores=8)

# plot toys
toyplots <- toys %>% group_by(toy_category) %>% filter(row_number()==1) %>%
  {split(., seq_len(nrow(.)))} %>% parallel::mclapply(function(row) dynplot::plot_strip_connections(row$gs[[1]], row$toy[[1]]), mc.cores = 8)

write_rds(toys, derived_file("toys.rds"))
```

```{r}
toys <- read_rds(derived_file("toys.rds"))

# get the scores when comparing the gs to toy
compare_toy <- function(gs, toy, id=toy$id) {
  scores <- dyneval:::calculate_metrics(gs, toy, metrics=metrics)$summary
  scores %>% mutate(toy_id=id)
}

# get the scores when both comparing gs to gs and comparing gs to toy
compare_gs_toy <- function(gs, toy, gs_grouped) {
  bind_rows(
    compare_toy(gs, toy) %>% mutate(comparison="gs-toy"),
    compare_toy(gs, gs, id=toy$id) %>% mutate(comparison="gs-gs"),
    compare_toy(gs_grouped, toy) %>% mutate(comparison="gs_groupid-toy")
  )
}

metrics <- c("mean_R_nx", "auc_R_nx", "Q_local", "Q_global", "correlation", "isomorphic", "node_edit_score", "node_edge_edit_score", "mantel_pval", "ged")

library(PRISM)

qsub_run(
  function(x) {
    library(tidyverse)
    parallel::mclapply(seq_len(nrow(toys)), function(row_id) {
      row <- dynutils::extract_row_to_list(toys, row_id)
      compare_gs_toy(.$gs, .$toy, .$gs_grouped)
    })
  },
  qsub_config = override_qsub_config(num_cores = 24),
  qsub_environment = list2env(list(compare_toy, compare_gs_toy, toys)
))


scores <- results %>% select(-starts_with("time"))

write_rds(scores, result_file("scores.rds"))
```

```{r}
timings <- results %>% 
  select(starts_with("time")) %>% 
  rename_all(list(function(x) substring(x, 6))) %>% 
  gather(metric_group_id, time) %>% 
  group_by(metric_group_id) %>% 
  summarise(time = mean(time))

metric_groups <- c("mean_R_nx"="coranking", "auc_R_nx"="coranking", "Q_local"="coranking", "Q_global"="coranking", "correlation"="correlation", "isomorphic"="isomorphic", "robbie_network_score"="robbie_network_score", "mantel_pval"="mantel", "ged"="ged") %>%
  enframe("metric_id", "metric_group_id")

timings %>% 
  left_join(metric_groups, by="metric_group_id") %>% 
  write_rds(result_file("timings.rds"))
```
